requirements.txt
torch
torchvision
timm
streamlit
pandas
numpy
scikit-learn
joblib
Pillow
